refactor: Repository interface をdomainに移動
- 責務が逆転しているため
- repositoryの実装クラスはdataのまま
