+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/typhoon/list/TyphoonListViewModel.kt
@@ -2,7 +2,7 @@ package com.yaeyama.linerchecker.ui.typhoon.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.yaeyama.linerchecker.repository.TyphoonRepository
+import com.yaeyama_liner_checker.domain.repository.TyphoonRepository
 import com.yaeyama_liner_checker.domain.tyhoon.Typhoon
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharingStarted.Companion.WhileSubscribed
