+++ data/src/main/kotlin/com/yaeyama/linerchecker/repository/TopStatusRepositoryImpl.kt
@@ -5,6 +5,7 @@ import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ktx.getValue
 import com.yaeyama.linerchecker.ext.reference
 import com.yaeyama.linerchecker.ext.valueEvents
+import com.yaeyama_liner_checker.domain.repository.TopStatusRepository
 import com.yaeyama_liner_checker.domain.top.Ports
 import com.yaeyama_liner_checker.domain.top.TopPort
 import kotlinx.coroutines.flow.Flow
