+++ data/src/main/kotlin/com/yaeyama/linerchecker/repository/TyphoonRepositoryImpl.kt
@@ -5,6 +5,7 @@ import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ktx.getValue
 import com.yaeyama.linerchecker.ext.reference
 import com.yaeyama.linerchecker.ext.valueEvents
+import com.yaeyama_liner_checker.domain.repository.TyphoonRepository
 import com.yaeyama_liner_checker.domain.tyhoon.Typhoon
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
