+++ domain/build.gradle.kts
@@ -1,3 +1,7 @@
 plugins {
     id("yaeyama.kotlin.library")
 }
+
+dependencies {
+    implementation(libs.coroutines.core)
+}
