+++ domain/src/main/java/com/yaeyama_liner_checker/domain/repository/TyphoonRepository.kt
@@ -1,4 +1,4 @@
-package com.yaeyama.linerchecker.repository
+package com.yaeyama_liner_checker.domain.repository
 
 import com.yaeyama_liner_checker.domain.tyhoon.Typhoon
 import kotlinx.coroutines.flow.Flow
