+++ domain/src/main/java/com/yaeyama_liner_checker/domain/repository/WeatherRepository.kt
@@ -1,5 +1,6 @@
-package com.yaeyama.linerchecker.repository
+package com.yaeyama_liner_checker.domain.repository
 
+import com.yaeyama_liner_checker.domain.common.UiState
 import com.yaeyama_liner_checker.domain.weather.WeatherInfo
 import kotlinx.coroutines.flow.Flow
